requirements.txt
opencv-python == 4.10.0.82
numpy == 1.23.4
Pillow == 9.2.0